Add location editor code
It's named LocationParentEditor because the tool started out as just a
quick way to edit the hierarchies, but as I completed more of the work,
I needed more of the stuff I had used in one tool.

The tool now has these major capabilities:

1. Edit location hierarchies
2. Edit cell-location associations
3. Edit location keywords

This tool DOES NOT edit ESM and ESP files directly. It dumps its output
in a form that's easy to write xEdit scripts against.