Add xEdit scripts
These scripts were used to process the locations when editing the
location hierarchy and the keywords associated with each location.

I was learning xEdit scripting at the same time that I was writing
these, so I'm sure there are better ways to do many of the things that
these do, but they achieved what I was trying to achieve with them at
the time.