Remove obsolete buttons
This commit removes the "Load hierarchy" and "Load cell mappings"
buttons since there's really no reason to load this data separately
anymore.

It also renames "Hide parents" to "Hide locs with parents" to avoid
ambiguity in the meaning of "parented". I've seen that used to mean
"make something a parent" or to mean "make something a child". The
meaning was the latter, but it was ambiguous.

Also updated the README for these changes.